Initial split preparations for #35713
