+++ homeassistant/components/plugwise/strings.json
@@ -12,6 +12,13 @@
   "config": {
     "step": {
       "user": {
+        "title": "Plugwise type",
+        "description": "Product:",
+        "data": {
+          "flow_type": "Connection type"
+        }
+      },
+      "user_gateway": {
         "title": "Connect to the Smile",
         "description": "Please enter:",
         "data": {
